+++ lib/util/priority_layer.dart
@@ -1,10 +1,10 @@
 class LayerPriority {
   static const int BACKGROUND = 10;
   static const int MAP = 20;
-  static const int COMPONENTS = 30;
+  static const int _COMPONENTS = 30;
 
   static int getComponentPriority(int bottom) {
-    return COMPONENTS + bottom;
+    return _COMPONENTS + bottom;
   }
 
   static int getAbovePriority(int highestPriority) {